1.4 adding more bonuses + some fixed functions from previous parts

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: vhyl <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/09 18:52:30 by vhyl              #+#    #+#             */
-/*   Updated: 2023/01/13 14:58:42 by vhyl             ###   ########.fr       */
+/*   Updated: 2023/01/15 14:24:40 by vhyl             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
